--- app/email.py
@@ -1,12 +1,0 @@
-from flask_mail import Mail, Message
-
-from app import app, mail
-
-def send_email(to, subject, template):
-    msg = Message(
-        subject,
-        recipients=[to],
-        html=template,
-        sender=app.config['MAIL_DEFAULT_SENDER']
-    )
-    mail.send(msg)
